tutorials-April2026-Batch/Session 2 - Personal Researcher/requirements.txt
beautifulsoup4
google-genai
elevenlabs
wikipedia
python-dotenv
requests
rich
streamlit
